Unite record completion logic into a single module
